build fix

+++ core/block_crypt.cpp
@@ -908,7 +908,7 @@ namespace beam
 			<< StartDifficulty.m_Packed
 			<< Block::PoW::K
 			<< Block::PoW::N
-			<< Block::PoW::NonceType::nBits
+			<< (uint32_t) Block::PoW::NonceType::nBits
 			<< uint32_t(0) // increment this whenever we change something in the protocol
 			// out
 			>> Checksum;
